+++ R/executeChecks.R
@@ -24,7 +24,7 @@
 #' @param checks the checks to be executed, by default the missing values, the
 #' exposure duration and the quantity. Possible options are "missing",
 #' "exposureDuration", "type", "route", "sourceConcept", "daysSupply",
-#' "verbatimEndDate", "dose", "sig", "quantity" and "diagnosticsSummary"
+#' "verbatimEndDate", "dose", "sig", "quantity", "daysBetween" and "diagnosticsSummary"
 #' @param minCellCount minimum number of events to report- results
 #' lower than this will be obscured. If 0 all results will be reported.
 #' @param sample the number of samples, default 10.000
@@ -36,6 +36,11 @@
 #' @param earliestStartDate the earliest date from which a record can be included
 #' @param verbose verbose, default FALSE
 #' @param byConcept boolean argument whether to return results by Concept or overall only
+#' @param exposureTypeId id of the drug exposure type to be filtered on (e.g. only prescribed).
+#' By default all record types will be taken into account.
+#' @param outputFolder folder to write to. If NULL, results will not be written to file
+#' @param databaseId database identifier
+#' @param filename output file name, if NULL it will be equal to databaseId
 #'
 #' @return named list with results
 #' @export
@@ -43,9 +48,9 @@
 #' @examples
 #' \dontrun{
 #' db <- DBI::dbConnect(" Your database connection here ")
-#' cdm <- CDMConnector::cdm_from_con(
+#' cdm <- CDMConnector::cdmFromCon(
 #'   con = db,
-#'   cdm_schema = "cdm schema name"
+#'   cdmSchema = "cdm schema name"
 #' )
 #' result <- executeChecks(
 #'   cdm = cdm,
@@ -61,17 +66,28 @@ executeChecks <- function(cdm,
                           tablePrefix = NULL,
                           earliestStartDate = "2010-01-01",
                           verbose = FALSE,
-                          byConcept = TRUE) {
+                          byConcept = TRUE,
+                          exposureTypeId = NULL,
+                          outputFolder = NULL,
+                          databaseId = CDMConnector::cdmName(cdm),
+                          filename = NULL) {
   errorMessage <- checkmate::makeAssertCollection()
   checkDbType(cdm = cdm, type = "cdm_reference", messageStore = errorMessage)
   checkmate::assertNumeric(ingredients, min.len = 1, add = errorMessage)
   checkmate::assertNumeric(subsetToConceptId, add = errorMessage, null.ok = TRUE)
+  validateChecks(checks, messageStore = errorMessage)
   checkmate::assertTRUE(is.numeric(minCellCount), add = errorMessage)
   checkmate::assertNumeric(sample, len = 1, add = errorMessage, null.ok = TRUE)
   checkSampleMinCellCount(sample, minCellCount, messageStore = errorMessage)
   checkmate::assertCharacter(tablePrefix, len = 1, add = errorMessage, null.ok = TRUE)
   checkmate::assertDate(as.Date(earliestStartDate), add = errorMessage, null.ok = FALSE)
   checkLogical(verbose, messageStore = errorMessage)
+  checkLogical(byConcept, messageStore = errorMessage)
+  if (!is.null(outputFolder)) {
+    checkmate::assertPathForOutput(outputFolder, overwrite = TRUE, add = errorMessage)
+  }
+  checkmate::assertCharacter(databaseId, len = 1, add = errorMessage)
+  checkmate::assertCharacter(filename, len = 1, null.ok = TRUE, add = errorMessage)
   checkmate::reportAssertions(collection = errorMessage)
 
   resultList <- vector(mode = "list", length = length(ingredients))
@@ -90,8 +106,17 @@ executeChecks <- function(cdm,
           tablePrefix = tablePrefix,
           earliestStartDate = earliestStartDate,
           verbose = verbose,
-          byConcept = byConcept
+          byConcept = byConcept,
+          exposureTypeId = exposureTypeId
         )
+
+        # write result
+        if (!is.null(outputFolder)) {
+          writeIngredientResultToDisk(resultList = ingredientResult,
+                                      databaseId = databaseId,
+                                      outputFolder = outputFolder,
+                                      clearDBDir = (i == 1))
+        }
       },
       error = function(e) {
         warning(e)
@@ -106,6 +131,9 @@ executeChecks <- function(cdm,
     dplyr::mutate(package_version = as.character(utils::packageVersion("DrugExposureDiagnostics")))
   resultList <- append(resultList, list("metadata" = metaData))
 
+  if (!is.null(outputFolder)) {
+    writeZipToDisk(metaData, databaseId, outputFolder, filename)
+  }
   return(resultList)
 }
 
@@ -129,6 +157,8 @@ executeChecks <- function(cdm,
 #' @param earliestStartDate the earliest date from which a record can be included
 #' @param verbose verbose, default FALSE
 #' @param byConcept boolean argument whether to return results by Concept or overall only
+#' @param exposureTypeId id of the drug exposure type to be filtered on (e.g. only prescribed).
+#' By default all record types will be taken into account.
 #'
 #' @return named list with results
 executeChecksSingleIngredient <- function(cdm,
@@ -140,7 +170,8 @@ executeChecksSingleIngredient <- function(cdm,
                                           tablePrefix = NULL,
                                           earliestStartDate = "2010-01-01",
                                           verbose = FALSE,
-                                          byConcept = FALSE) {
+                                          byConcept = FALSE,
+                                          exposureTypeId = NULL) {
   errorMessage <- checkmate::makeAssertCollection()
   checkDbType(cdm = cdm, type = "cdm_reference", messageStore = errorMessage)
   checkIsIngredient(cdm = cdm, conceptId = ingredient, messageStore = errorMessage)
@@ -186,24 +217,13 @@ executeChecksSingleIngredient <- function(cdm,
   }
 
   if (verbose == TRUE) {
-    start <- printDurationAndMessage("Progress: getting drug strength for ingredient", start)
-  }
-
-  cdm[["ingredient_drug_strength"]] <- getDrugStrength(
-    cdm = cdm,
-    ingredient = ingredient,
-    includedConceptsTable = "ingredient_concepts",
-    tablePrefix = tablePrefix
-  ) %>%
-    dplyr::compute(name = "ingredient_drug_strength")
-
-  if (verbose == TRUE) {
     start <- printDurationAndMessage("Progress: getting drug records for ingredient", start)
   }
   cdm[["ingredient_drug_records"]] <- getDrugRecords(
     cdm = cdm,
     ingredient = ingredient,
     includedConceptsTable = "ingredient_concepts",
+    exposureTypeId = exposureTypeId,
     tablePrefix = tablePrefix
   ) %>%
     dplyr::compute(name = "ingredient_drug_records")
@@ -236,14 +256,20 @@ executeChecksSingleIngredient <- function(cdm,
     if (verbose == TRUE) {
       start <- printDurationAndMessage("Progress: sampling drug records", start)
     }
-    if (dplyr::pull(dplyr::tally(dplyr::filter(cdm[["ingredient_drug_records"]], .data$drug_exposure_start_date > .env$earliestStartDate)), .data$n) < sampleSize) {
-      message("population after earliestStartDate smaller than sample, sampling ignored")
-      cdm[["ingredient_drug_records"]] <- cdm[["ingredient_drug_records"]] %>%
-        dplyr::filter(.data$drug_exposure_start_date > .env$earliestStartDate) %>%
-        dplyr::compute()
+    if (!is.null(earliestStartDate)) {
+      if (dplyr::pull(dplyr::tally(dplyr::filter(cdm[["ingredient_drug_records"]], .data$drug_exposure_start_date > .env$earliestStartDate)), .data$n) < sampleSize) {
+        message("population after earliestStartDate smaller than sample, sampling ignored")
+        cdm[["ingredient_drug_records"]] <- cdm[["ingredient_drug_records"]] %>%
+          dplyr::filter(.data$drug_exposure_start_date > .env$earliestStartDate) %>%
+          dplyr::compute()
+      } else {
+        cdm[["ingredient_drug_records"]] <- cdm[["ingredient_drug_records"]] %>%
+          dplyr::filter(.data$drug_exposure_start_date > .env$earliestStartDate) %>%
+          dplyr::slice_sample(n = sampleSize) %>%
+          dplyr::compute()
+      }
     } else {
       cdm[["ingredient_drug_records"]] <- cdm[["ingredient_drug_records"]] %>%
-        dplyr::filter(.data$drug_exposure_start_date > .env$earliestStartDate) %>%
         dplyr::slice_sample(n = sampleSize) %>%
         dplyr::compute()
     }
@@ -388,6 +414,18 @@ executeChecksSingleIngredient <- function(cdm,
     }
   }
 
+  drugTimeBetween <- drugTimeBetweenByConcept <- NULL
+  if ("daysBetween" %in% checks) {
+    if (verbose == TRUE) {
+      start <- printDurationAndMessage("Progress: check time between drug records", start)
+    }
+
+    drugTimeBetween <- summariseTimeBetween(cdm, "ingredient_drug_records", byConcept = FALSE, sampleSize = sampleSize) %>% dplyr::collect()
+    if (isTRUE(byConcept)) {
+      drugTimeBetweenByConcept <- summariseTimeBetween(cdm, "ingredient_drug_records", byConcept = byConcept, sampleSize = sampleSize) %>% dplyr::collect()
+    }
+  }
+
   if (!is.null(tablePrefix)) {
     if (verbose == TRUE) {
       start <- printDurationAndMessage("Cleaning up tables", start)
@@ -422,7 +460,9 @@ executeChecksSingleIngredient <- function(cdm,
     "drugSig" = drugSig,
     "drugSigByConcept" = drugSigByConcept,
     "drugQuantity" = drugQuantity,
-    "drugQuantityByConcept" = drugQuantityByConcept
+    "drugQuantityByConcept" = drugQuantityByConcept,
+    "drugTimeBetween" = drugTimeBetween,
+    "drugTimeBetweenByConcept" = drugTimeBetweenByConcept
   )
 
   # add summary table