+++ tests/testthat/test-SyntheaSqlServer.R
@@ -12,7 +12,9 @@ test_that("test methods against test server", {
     PWD      = Sys.getenv("TESTDB_PWD"),
     Port     = Sys.getenv("TESTDB_PORT")
   )
-  cdm <- CDMConnector::cdm_from_con(con, cdm_schema = Sys.getenv("TESTDB_CDM_SCHEMA"))
+  cdm <- CDMConnector::cdmFromCon(con,
+                                  cdmSchema = Sys.getenv("TESTDB_CDM_SCHEMA"),
+                                  writeSchema = Sys.getenv("TESTDB_WRITE_SCHEMA"))
 
   result <- executeChecks(cdm = cdm, ingredients = c(1125315), verbose = TRUE) # acetaminophen