+++ NEWS.md
@@ -1,5 +1,24 @@
 # DrugExposureDiagnostics (development version)
 
+# DrugExposureDiagnostics 1.1.3
+* Add time in between check
+* Integration with Darwin shiny modules
+
+# DrugExposureDiagnostics 1.1.2
+* Add option to run for an exposure type
+* Fix rounding issue drug_type_concept_id
+* Add error if wrong check is passed
+* Fix earliestStartDate NULL behaviour
+* Move shiny to suggests
+
+# DrugExposureDiagnostics 1.1.1
+* Update shiny app: fix NA in boxplots and add dropdown drugs-missing
+* CDMConnector v2 changes
+* Add shiny modules
+
+# DrugExposureDiagnostics 1.1.0
+* Add params to executeChecks so it will write results to disk, no need to separately call writeResultToDisk
+
 # DrugExposureDiagnostics 1.0.10
 * Fix sample is null bug
 * Fix too many rows in diagnostics summary